+++ Pod/Classes/KRGeoHex3.h
@@ -7,20 +7,21 @@
 //
 
 #import <Foundation/Foundation.h>
+#import <stdint.h>
 
 typedef NSDictionary<NSString *,NSNumber *> geohex3_location_dict_t;
 typedef NSArray<geohex3_location_dict_t *>  geohex3_locations_t;
 
 @interface KRGeoHex3 : NSObject
 
-- (nonnull instancetype)initWithX:(const NSInteger)x y:(const NSInteger)y level:(const NSUInteger)level;
+- (nonnull instancetype)initWithX:(const int64_t)x y:(const int64_t)y level:(const NSUInteger)level;
 - (nonnull instancetype)initWithLat:(const double)lat lng:(const double)lng level:(const NSUInteger)level;
 - (nonnull instancetype)initWithCode:(nonnull const NSString *)code;
 - (NSUInteger)getLevel;
 - (double)getSize;
 - (nonnull NSString *)getCode;
-- (NSInteger)getX;
-- (NSInteger)getY;
+- (int64_t)getX;
+- (int64_t)getY;
 - (double)getLat;
 - (double)getLng;
 - (nonnull geohex3_locations_t *)getPolygon;
